requirements.txt
python-telegram-bot>=21.0
python-dotenv==1.0.0
sqlalchemy>=2.0.36
psycopg[binary]>=3.1.0
python-dateutil==2.8.2

